+++ internal/shared/shared.go
@@ -31,13 +31,6 @@ func WithCacheController(cacheController cachecontroller.CacheController) Shared
 	}
 }
 
-// WithShadowCacheController allows overriding the default shadow cacheController created in NewSharedDatastoreResources().
-func WithShadowCacheController(cacheController cachecontroller.CacheController) SharedDatastoreResourcesOpt {
-	return func(scr *SharedDatastoreResources) {
-		scr.ShadowCacheController = cacheController
-	}
-}
-
 // SharedDatastoreResources contains resources that can be shared across Check requests.
 type SharedDatastoreResources struct {
 	SingleflightGroup     *singleflight.Group
@@ -45,8 +38,6 @@ type SharedDatastoreResources struct {
 	ServerCtx             context.Context
 	CheckCache            storage.InMemoryCache[any]
 	CacheController       cachecontroller.CacheController
-	ShadowCheckCache      storage.InMemoryCache[any]
-	ShadowCacheController cachecontroller.CacheController
 	Logger                logger.Logger
 	SharedIteratorStorage *sharediterator.Storage
 }
@@ -83,27 +74,6 @@ func NewSharedDatastoreResources(
 		s.CacheController = cachecontroller.NewCacheController(ds, s.CheckCache, settings.CacheControllerTTL, settings.CheckIteratorCacheTTL, cachecontroller.WithLogger(s.Logger))
 	}
 
-	// The default behavior is to use the same cache instance for both the
-	// check cache and the shadow cache. However, if the user opts in to use a
-	// separate cache instance for the shadow cache, we need to create new
-	// instances.
-	s.ShadowCheckCache = s.CheckCache
-	s.ShadowCacheController = s.CacheController
-
-	if settings.ShouldCreateShadowNewCache() {
-		var err error
-		s.ShadowCheckCache, err = storage.NewInMemoryLRUCache([]storage.InMemoryLRUCacheOpt[any]{
-			storage.WithMaxCacheSize[any](int64(settings.CheckCacheLimit)),
-		}...)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	if settings.ShouldCreateShadowCacheController() {
-		s.ShadowCacheController = cachecontroller.NewCacheController(ds, s.ShadowCheckCache, settings.CacheControllerTTL, settings.CheckIteratorCacheTTL, cachecontroller.WithLogger(s.Logger))
-	}
-
 	for _, opt := range opts {
 		opt(s)
 	}
@@ -117,8 +87,5 @@ func (s *SharedDatastoreResources) Close() {
 	s.WaitGroup.Wait()
 	if s.CheckCache != nil {
 		s.CheckCache.Stop()
-	}
-	if s.ShadowCheckCache != nil && s.CheckCache != s.ShadowCheckCache {
-		s.ShadowCheckCache.Stop()
 	}
 }