+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Try to keep listed changes to a concise bulleted list of simple explanations of changes. Aim for the amount of information needed so that readers can understand where they would look in the codebase to investigate the changes' implementation, or where they would look in the documentation to understand how to make use of the change in practice - better yet, link directly to the docs and provide detailed information there. Only elaborate if doing so is required to avoid breaking changes or experimental features from ruining someone's day.
 
 ## [Unreleased]
+### Changed
+- Breaking: Replace shadow check resolver with shadow check command. [#2698](https://github.com/openfga/openfga/pull/2698)
+  > The following breaking changes are related to the `server.Server` interface. If you are not customizing the server, then these changes should not impact your usage of OpenFGA.
+    - Removed `WithShadowListObjectsCheckResolverEnabled` server option.
+    - Removed `WithShadowListObjectsCheckResolverTimeout` server option.
+    - Removed `WithShadowListObjectsCheckResolverSamplePercentage` server option.
+  > The following breaking changes are related to the `CheckResolver`. If you are not customizing a check resolver, then these changes should not impact your usage of OpenFGA.
+    - Removed `WithLocalShadowCheckerOpts` option.
+    - Removed `WithShadowResolverEnabled` option.
+    - Removed `WithShadowResolverOpts` option.
+    - Removed `ShadowResolver` struct implementation, replaced by `ShadowCheckCommand`.
+
 ### Fixed
 - Revert spf13/viper back to v.1.20.1 to avoid bumping sourcegraph/conc to an unreleased version as it causes performance degradation. [#2706](https://github.com/openfga/openfga/pull/2706)